# api-documentation/api/swagger.yaml
swagger: "2.0"
info:
  description: "This server takes in a number and outputs a roman numeral. Extension 1: Expand the set of numbers to 1-3999. Extension 2: Expand the set of numbers to include the Vinculum addition using UTF-8 to roman numerals greater than 3999 and handle all numbers up to 2,200,000,000."
  version: "1.0.0"
  title: "Decimal number to roman number converter"
basePath: "/"
tags:
- name: "romannumeral"
  description: "Number to roman number notation conversion"
- name: "monitoring"
  description: "Check server health"
schemes:
- "https"
- "http"
paths:
  /romannumeral:
    get:
      tags:
      - "romannumeral"
      summary: "Add a new pet to the store"
      description: ""
      operationId: "convertNumToRoman"
      produces:
      - "application/json"
      parameters:
      - name: "query"
        in: "query"
        description: "(Decimal) number to convert to roman number"
        required: true
        type: "integer"
      responses:
        200:
          description: "Converted roman number"
          schema:
            $ref: "#/definitions/RomanNumber"
        422:
          description: "Invalid input - The request was well-formed but was unable to be followed due to semantic errors."
          schema:
            $ref: "#/definitions/ConversionError"
          
  /heartbeat:
    get:
      tags:
      - "monitoring"
      summary: "Server heartbeat"
      description: "Ask for a heartbeat of the server (respond 200 if up and running)"
      operationId: "heartbeat"
      produces:
      - "application/json"
      responses:
        200:
          description: "Server is healthy"
          
  /version:
    get:
      tags:
      - "monitoring"
      summary: "API version"
      description: "Gets the API version"
      operationId: "version"
      produces:
      - "application/json"
      responses:
        200:
          description: "Server is healthy"
          schema:
            $ref: "#/definitions/APIInfo"
          
definitions:
  APIInfo:
    type: "object"
    properties:
      version:
        type: "string"
        description: "semantic version number of the API"
  RomanNumber:
    type: "object"
    properties:
      roman:
        type: "string"
        description: "roman number (in UTF-8 format)"
  ConversionError:
    type: "object"
    properties:
      error:
        type: "string"
        description: "error code"
      message:
        type: "string"
        description: "error message"
      apiVersion:
        type: "string"
        description: "api version"
      status:
        type: "integer"
        description: "HTTP status code"
  